requirements.txt
Flask
gunicorn
requests
google-generativeai
nltk
joblib
pandas
scikit-learn
